refactor: centralize trading logic
